Move Mappers Test And Samples
